Change amount field in MpesaPaymentSerializer from DecimalField to IntegerField

+++ api/serializers.py
@@ -2,7 +2,7 @@ from rest_framework import serializers
 
 class MpesaPaymentSerializer(serializers.Serializer):
     phone_number = serializers.CharField(max_length=15)
-    amount = serializers.DecimalField(max_digits=10, decimal_places=2)
+    amount = serializers.IntegerField()
 
 class MpesaCallbackSerializer(serializers.Serializer):
     Body = serializers.DictField()
